# openapi/p4pa-debt-position.openapi.yaml
openapi: 3.0.1
info:
  title: p4pa-debt-positions
  description: Api and Models
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /debt-positions:
    post:
      tags:
        - debt-position
      operationId: createDebtPosition
      parameters:
        - name: massive
          in: query
          description: Indicates if the operation is massive (true) or single (false).
          required: false
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtPositionDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "500":
          description: Internal server error
  /debt-positions/{debtPositionId}:
    get:
      tags:
        - debt-position
      operationId: getDebtPosition
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
    delete:
      tags:
        - debt-position
      operationId: deleteDebtPosition
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      responses:
        "204":
          description: No content
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/by-installmentId/{installmentId}:
    get:
      tags:
        - debt-position
      operationId: getDebtPositionByInstallmentId
      parameters:
        - name: installmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the installment
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/by-iuv/{organizationId}/{iuv}:
    get:
      tags:
        - debt-position
      operationId: getDebtPositionsByOrganizationIdAndIuv
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the organization
        - name: iuv
          in: path
          required: true
          schema:
            type: string
          description: The iuv of the installments
        - name: debtPositionOrigin
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/DebtPositionOrigin"
            description: An array of debt position origins to filter by; if null, no filter is applied
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/by-iud/{organizationId}/{iud}:
    get:
      tags:
        - debt-position
      operationId: getDebtPositionsByOrganizationIdAndIud
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the organization
        - name: iud
          in: path
          required: true
          schema:
            type: string
          description: The iud of the installments
        - name: debtPositionOrigin
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/DebtPositionOrigin"
            description: An array of debt position origins to filter by; if null, no filter is applied
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/{debtPositionId}/publish:
    put:
      tags:
        - debt-position
      operationId: publishDebtPosition
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionDTO"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/ingestion-flow-file/{ingestionFlowFileId}:
    get:
      tags:
        - debt-position
      operationId: getDebtPositionsByIngestionFlowFileId
      parameters:
        - name: ingestionFlowFileId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the ingestion flow file
        - name: statusToExclude
          in: query
          description: Indicates the list of status to be excluded
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/InstallmentStatus"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDebtPositions"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
      x-spring-paginated: true
...
  /debt-positions/{debtPositionId}/finalize-sync-status:
    put:
      tags:
        - debt-position
      operationId: finalizeSyncStatus
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncStatusUpdateRequestDTO"
        required: true
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionDTO'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/{debtPositionId}/check-installment-expiration:
    put:
      tags:
        - debt-position
      operationId: checkAndUpdateInstallmentExpiration
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionDTO'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/installment-synchronize:
    put:
      tags:
        - debt-position
      operationId: installmentSynchronize
      parameters:
        - name: origin
          in: query
          description: Indicates the origin of debt position to synchronize
          required: true
          schema:
            $ref: "#/components/schemas/DebtPositionOrigin"
        - name: massive
          in: query
          description: Indicates if the operation is massive (true) or single (false).
          required: false
          schema:
            type: boolean
            default: false
        - name: partialChange
          in: query
          description: Indicates if the operation is a cumulative operation (true), which at the end of the process will invoke again the API with the parameter set to false, or complete (false).
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallmentSynchronizeDTO"
        required: true
      responses:
        "201":
          description: Ok
          headers:
            x-workflow-id:
              schema:
                type: string
              description: The workflow id created.
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/update-installment-notification-date:
    put:
      tags:
        - debt-position
      operationId: updateInstallmentNotificationDate
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                debtPositionId:
                  type: integer
                  format: int64
                notificationDate:
                  type: string
                  format: date-time
                nav:
                  type: array
                  items:
                    type: string
        required: true
      responses:
        "201":
          description: Ok
          headers:
            x-workflow-id:
              schema:
                type: string
              description: The workflow id created.
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/{debtPositionId}/manage-installments:
    put:
      tags:
        - debt-position
      operationId: manageDebtPositionInstallments
      parameters:
        - name: debtPositionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The ID of the debt position
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManageDebtPositionDTO"
        required: true
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionDTO'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
  /debt-positions/update-notification-fee:
    put:
      tags:
        - debt-position
      operationId: updateInstallmentNotificationFee
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationId:
                  type: integer
                  format: int64
                nav:
                  type: string
                newFeeCents:
                  type: integer
                  format: int64
              required:
                - organizationId
                - nav
                - newFeeCents
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallmentDTO'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPositionErrorDTO"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtPositionErrorDTO'
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    DebtPositionDTO:
      required:
        - debtPositionOrigin
        - organizationId
        - debtPositionTypeOrgId
        - paymentOptions
        - flagPuPagoPaPayment
      type: object
      x-implements: ["it.gov.pagopa.pu.debtpositions.dto.BaseDebtPosition"]
      properties:
        debtPositionId:
          type: integer
          format: int64
        iupdOrg:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/DebtPositionStatus"
        debtPositionOrigin:
          $ref: "#/components/schemas/DebtPositionOrigin"
        organizationId:
          type: integer
          format: int64
        debtPositionTypeOrgId:
          type: integer
          format: int64
        validityDate:
          type: string
          format: date
        flagIuvVolatile:
          type: boolean
        multiDebtor:
          type: boolean
        flagPuPagoPaPayment:
          type: boolean
        creationDate:
          type: string
          format: date-time
        updateDate:
          type: string
          format: date-time
        updateOperatorExternalId:
          type: string
        updateTraceId:
          type: string
        paymentOptions:
          type: array
          items:
            $ref: "#/components/schemas/PaymentOptionDTO"
...
    DebtPositionStatus:
      type: string
      enum:
        - TO_SYNC
        - REPORTED
        - PAID
        - PARTIALLY_PAID
        - CANCELLED
        - EXPIRED
        - UNPAID
        - DRAFT
    DebtPositionOrigin:
      type: string
      enum:
        - ORDINARY
        - ORDINARY_SIL
        - SPONTANEOUS
        - SPONTANEOUS_SIL
        - SECONDARY_ORG
        - RECEIPT_FILE
        - RECEIPT_PAGOPA
        - REPORTING_PAGOPA
...
    InstallmentStatus:
      type: string
      enum:
        - TO_SYNC
        - REPORTED
        - PAID
        - CANCELLED
        - INVALID
        - EXPIRED
        - UNPAID
        - UNPAYABLE
        - DRAFT
...
    SyncStatusUpdateRequestDTO:
      type: object
      required:
        - iupd2finalize
        - iupdSyncError
      properties:
        iupd2finalize:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/SyncCompleteDTO"
        iupdSyncError:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/SyncErrorDTO"
    SyncCompleteDTO:
      type: object
      required:
        - newStatus
      properties:
        newStatus:
          $ref: "#/components/schemas/InstallmentStatus"
...
    DebtPositionErrorDTO:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - DEBT_POSITION_BAD_REQUEST
            - DEBT_POSITION_CONFLICT
            - DEBT_POSITION_FORBIDDEN
            - DEBT_POSITION_GENERIC_ERROR
            - DEBT_POSITION_NOT_FOUND
        message:
          type: string
...
    PagedDebtPositions:
      type: object
      required:
        - content
        - size
        - totalElements
        - totalPages
        - number
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/DebtPositionDTO"
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...